requirements.txt
django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
django-environ==0.4.5
pymysql==1.0.2